feat: implementando validações usuario

+++ src/modules/users/users.service.ts
@@ -1,4 +1,8 @@
-import { Injectable, ConflictException } from '@nestjs/common';
+import {
+  Injectable,
+  ConflictException,
+  NotFoundException,
+} from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { UserRepository } from './repositories/users.repository';
@@ -33,14 +37,19 @@ export class UsersService {
 
   async findOne(id: number) {
     const user = await this.repository.findOne(id);
+    if (!user) throw new NotFoundException(`User ${id} is not found`);
     return user;
   }
 
   async update(id: number, updateUserDto: UpdateUserDto) {
+    const user = await this.repository.findOne(id);
+    if (!user) throw new NotFoundException(`User ${id} is not found`);
     return await this.repository.update(id, updateUserDto);
   }
 
   async remove(id: number) {
+    const user = await this.repository.findOne(id);
+    if (!user) throw new NotFoundException(`User ${id} is not found`);
     await this.repository.remove(id);
   }
 }
